Move prelogin handshake decoder to its own class
